feat(events): promote claimEvent to a shared idempotency primitive
At-least-once delivery means consumers that create non-idempotent side
effects must dedupe on the event id. notifications already had this
(processed_events + claimEvent); lift it into @adopt-dont-shop/events as
the canonical, schema-agnostic primitive (table resolved via search_path)
so every service shares one implementation and new consumers adopt it by
adding the processed_events migration + calling claimEvent. notifications
now re-exports the shared helper (call sites unchanged). Naturally
idempotent consumers (GDPR erase, ON CONFLICT report inserts, deterministic
completion ids) are intentionally left as-is rather than gold-plated.
